+++ index.html
@@ -250,7 +250,7 @@
     </section>
   </template>
 
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
   <script src="js/function.js"></script>
 
 </body>
